Remove postgame encounters to fix sorting
