+++ src/models/Game.ts
@@ -3,10 +3,11 @@ import LocalName from "@/models/LocalName";
 
 export default interface Game {
     name: string;
+    iconURL: string;
     versionGroup: string;
     pokedex: LocalName[];
     segments: LocalSegment[];
     starterSlugs: string[];
     startingTownSlug: string;
-    iconURL: string;
+    postgameMethods: string[];
 }
